web/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
httpx==0.27.2
pydantic==2.10.3
pydantic-settings==2.6.1
itsdangerous==2.2.0
python-jose[cryptography]==3.3.0
aiosqlite==0.20.0
pyyaml==6.0.2
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
